functional tests: use pytext fixtures, delete utils.py
