+++ tests/functional/cassettes/test_delete_source_and_their_docs.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
     body: '{"username": "journalist", "passphrase": "correct horse battery staple
-      profanity oil chewy", "one_time_code": "177711"}'
+      profanity oil chewy", "one_time_code": "994892"}'
     headers:
       Accept:
       - '*/*'
