+++ tests/functional/cassettes/test_logout_as_journalist.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
     body: '{"username": "journalist", "passphrase": "correct horse battery staple
-      profanity oil chewy", "one_time_code": "333598"}'
+      profanity oil chewy", "one_time_code": "994892"}'
     headers:
       Accept:
       - '*/*'
