+++ tests/functional/cassettes/test_receive_message_from_source.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
     body: '{"username": "journalist", "passphrase": "correct horse battery staple
-      profanity oil chewy", "one_time_code": "795461"}'
+      profanity oil chewy", "one_time_code": "994892"}'
     headers:
       Accept:
       - '*/*'
