+++ tests/functional/test_delete_source.py
@@ -7,18 +7,15 @@ https://github.com/freedomofpress/securedrop-client/wiki/Test-plan#basic-client-
 import pytest
 from flaky import flaky
 from PyQt5.QtCore import Qt
-from .utils import get_safe_tempdir, get_logged_in_test_context
 
 
 @flaky
 @pytest.mark.vcr()
-def test_delete_source_and_their_docs(qtbot, mocker):
+def test_delete_source_and_their_docs(functional_test_logged_in_context, qtbot, mocker):
     """
     It's possible to delete a source and see it removed from the UI.
     """
-    totp = "177711"
-    tempdir = get_safe_tempdir()
-    gui, controller = get_logged_in_test_context(tempdir, qtbot, totp)
+    gui, controller, temmpdir = functional_test_logged_in_context
     qtbot.wait(1000)
 
     def check_for_sources():
