+++ tests/functional/test_send_reply.py
@@ -7,19 +7,16 @@ https://github.com/freedomofpress/securedrop-client/wiki/Test-plan#basic-client-
 import pytest
 from flaky import flaky
 from PyQt5.QtCore import Qt
-from .utils import get_safe_tempdir, get_logged_in_test_context
 
 
 @flaky
 @pytest.mark.vcr()
-def test_send_reply_to_source(qtbot, mocker):
+def test_send_reply_to_source(functional_test_logged_in_context, qtbot, mocker):
     """
     It's possible to send a reply to a source and see it show up in the
     conversation window.
     """
-    totp = "778326"
-    tempdir = get_safe_tempdir()
-    gui, controller = get_logged_in_test_context(tempdir, qtbot, totp)
+    gui, controller, tempdir = functional_test_logged_in_context
     qtbot.wait(1000)
 
     def check_for_sources():
